cleaning the code and adding more sprites

+++ Utils/ProjectileExtras.cs
@@ -8,16 +8,17 @@ using Terraria.ModLoader;
 namespace MagicTridents.Utils
 {
     /// <summary>
-    /// This class is to turn easier using Projectiles
+    /// This class is used to facilitate the usage of projectiles.
     /// </summary>
     public static class ProjectileExtras
     {
         /// <summary>
-        /// 
+        /// Applies gravity to the vertical movement of the projectile.
         /// </summary>
-        /// <param name="proj"></param>
-        /// <param name="power"></param>
-        /// <param name="limit"></param>
+        /// <param name="velocityY">The vertical velocity of the projectile.</param>
+        /// <param name="power">The strength of the gravity to be applied. The default value is 0.1f.</param>
+        /// <param name="limit">The maximum limit of the vertical velocity. The default value is 16.</param>
+        /// <returns>The vertical velocity of the projectile after applying gravity.</returns>
         public static float ApplyGravity(float velocityY, float power = 0.1f, float limit = 16)
         {
             if (velocityY > limit)
@@ -31,10 +32,10 @@ namespace MagicTridents.Utils
         }
 
         /// <summary>
-        /// 
+        /// Applies bouncing effect to the projectile.
         /// </summary>
-        /// <param name="proj"></param>
-        /// <param name="oldVelocity"></param>
+        /// <param name="proj">The projectile to be affected.</param>
+        /// <param name="oldVelocity">The previous velocity of the projectile.</param>
         public static void ApplyBounce(ModProjectile proj, Vector2 oldVelocity)
         {
             proj.Projectile.penetrate--;
@@ -47,19 +48,27 @@ namespace MagicTridents.Utils
                 Collision.HitTiles(proj.Projectile.position, proj.Projectile.velocity, proj.Projectile.width, proj.Projectile.height);
                 SoundEngine.PlaySound(SoundID.Item10, proj.Projectile.position);
 
-                // If the Projectile hits the left or right side of the tile, reverse the X velocity
+                // If the projectile hits the left or right side of the block, invert the velocity on the X axis
                 if (Math.Abs(proj.Projectile.velocity.X - oldVelocity.X) > float.Epsilon)
                 {
                     proj.Projectile.velocity.X = -oldVelocity.X;
                 }
 
-                // If the Projectile hits the top or bottom side of the tile, reverse the Y velocity
+                // If the projectile hits the top or bottom side of the block, invert the velocity on the Y axis
                 if (Math.Abs(proj.Projectile.velocity.Y - oldVelocity.Y) > float.Epsilon)
                 {
                     proj.Projectile.velocity.Y = -oldVelocity.Y;
                 }
             }
         }
+
+        /// <summary>
+        /// Applies orbiting around the player to the projectile.
+        /// </summary>
+        /// <param name="projectile">The projectile to be affected.</param>
+        /// <param name="player">The player around which the projectile will orbit.</param>
+        /// <param name="radius">The radius of the orbit.</param>
+        /// <param name="speed">The rotation speed of the projectile.</param>
         public static void ApplyOrbitingPlayer(ModProjectile projectile, Player player, float radius, float speed)
         {
             double deg = (double)projectile.Projectile.ai[1];
@@ -70,9 +79,17 @@ namespace MagicTridents.Utils
             projectile.Projectile.position.Y = player.Center.Y - (int)(Math.Sin(rad) * dist) - projectile.Projectile.height / 2;
             projectile.Projectile.ai[1] += speed;
         }
+
+        /// <summary>
+        /// Applies orbiting around any point to the projectile.
+        /// </summary>
+        /// <param name="projectile">The projectile to be affected.</param>
+        /// <param name="Center">The center around which the projectile will orbit.</param>
+        /// <param name="radius">The radius of the orbit.</param>
+        /// <param name="speed">The rotation speed of the projectile.</param>
         public static void ApplyOrbitingAnything(ModProjectile projectile, Vector2 Center, float radius, float speed)
         {
-            double deg = (double)projectile.Projectile.ai[1];
+            double deg = projectile.Projectile.ai[1];
             double rad = deg * (Math.PI / 180);
             double dist = radius;
